requirements.txt
catboost==1.2.8
genai==2.1.0
google-genai==1.16.1
google-generativeai==0.8.5
lightgbm==4.6.0
streamlit==1.46.1
xgboost==2.1.4